feat: add auto-focus on rejected submissions
the submit function of the ItemForm component has some validators. If
user gives a valid input and submits, the item gets added. If the user
does not give a valid input and submits, the unvalid item does not get
added. In item addition, the name input gets auto focused. Now this commit
adds a mechanism to make the name input focused after rejected submits too.